requirements.txt
google-genai
python-dotenv
yfinance
mcp
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
